WIP 0.2.0

+++ backend/backend_food_easy/src/main/java/com/food_easy_back/backend_food_easy/service/UserServiceImpl/UserServiceImpl.java
@@ -1,19 +1,30 @@
 package com.food_easy_back.backend_food_easy.service.UserServiceImpl;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.food_easy_back.backend_food_easy.model.dao.StoreDao;
 import com.food_easy_back.backend_food_easy.model.dao.UserDao;
+import com.food_easy_back.backend_food_easy.model.dto.OwnerCreateRequestDto;
+import com.food_easy_back.backend_food_easy.model.dto.StoreCreateRequestDto;
 import com.food_easy_back.backend_food_easy.model.dto.UserCreateRequestDto;
 import com.food_easy_back.backend_food_easy.model.dto.UserUpdateRequestDto;
 import com.food_easy_back.backend_food_easy.model.entity.StoreEntity;
 import com.food_easy_back.backend_food_easy.model.entity.UserEntity;
 import com.food_easy_back.backend_food_easy.model.entity.UserRoleEntity;
+import com.food_easy_back.backend_food_easy.service.IStoreService;
 import com.food_easy_back.backend_food_easy.service.IUserService;
 
 import jakarta.transaction.Transactional;
@@ -23,32 +34,87 @@ public class UserServiceImpl implements IUserService {
 
     private final UserDao userDao;
     private final StoreDao storeDao;
+    private final PasswordEncoder passwordEncoder;
+    private final IStoreService storeService;
 
     
     @Autowired
-    public UserServiceImpl(UserDao userDao, StoreDao storeDao) {
+    public UserServiceImpl(UserDao userDao, StoreDao storeDao, PasswordEncoder passwordEncoder,
+            IStoreService storeService) {
         this.userDao = userDao;
         this.storeDao = storeDao;
+        this.passwordEncoder = passwordEncoder;
+        this.storeService = storeService;
     }
 
 
+    //Metodo para guardar el usuario desde el owner
     @Transactional
     @Override
-    public UserEntity saveUser(UserCreateRequestDto userdto) {
+    public UserEntity saveUserByAdmin(UserCreateRequestDto userdto) {
 
-        StoreEntity storeEntity = storeDao.findById(userdto.getStore()).orElseThrow(null);
+        if(existUsername(userdto.getUsername())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El usuario ya existe" );
+        }else{
+            userdto.setPassword(passwordEncoder.encode(userdto.getPassword()));
+            UserEntity admin= findByUsername(getCurrentUsername());
+            Integer store= admin.getStore().getIdStore();
+            userdto.setStore(store);
+            StoreEntity storeEntity = storeDao.findById(userdto.getStore()).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Store no encontrado"));
+            UserEntity user = UserEntity.builder()
+                            .name(userdto.getName())
+                            .lastName(userdto.getLastname())
+                            .email(userdto.getEmail())
+                            .password(userdto.getPassword())
+                            .username(userdto.getUsername())
+                            .store(storeEntity)
+                            .disabled(false)
+                            .locked(false)
+                            .build();
+            List<UserRoleEntity> userRoles = userdto.getRoles().stream()
+                                                    .map(role -> {
+                                                        UserRoleEntity userRole = new UserRoleEntity();
+                                                        userRole.setUsername(user.getUsername());
+                                                        userRole.setRole(role);
+                                                        userRole.setGrantedDate(LocalDateTime.now());
+                                                        userRole.setUser(user); 
+                                                        return userRole;
+                                                    })
+                                                    .collect(Collectors.toList());
+
+            user.setRoles(userRoles);
+            return userDao.save(user);
+        }
+    }
+
+    //Metodo para guardar el usuario owner y el negocio desde el admin
+    @Transactional
+    @Override
+    public UserEntity saveOwner(OwnerCreateRequestDto requestDto) {
+        UserCreateRequestDto userDto = requestDto.getUser();
+        StoreCreateRequestDto storeDto = requestDto.getStore();
+        if(existUsername(userDto.getUsername())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El usuario ya existe" );
+        }
+
+        StoreEntity storeEntity = storeService.saveStore(storeDto);
+        userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
+        Integer store= storeEntity.getIdStore();
+        userDto.setStore(store);
+
+        userDto.setRoles(List.of("OWNER", "ADMIN", "USER"));
 
         UserEntity user = UserEntity.builder()
-                         .name(userdto.getName())
-                         .lastName(userdto.getLastname())
-                         .email(userdto.getEmail())
-                         .password(userdto.getPassword())
-                         .username(userdto.getUsername())
+                         .name(userDto.getName())
+                         .lastName(userDto.getLastname())
+                         .email(userDto.getEmail())
+                         .password(userDto.getPassword())
+                         .username(userDto.getUsername())
                          .store(storeEntity)
                          .disabled(false)
                          .locked(false)
                          .build();
-        List<UserRoleEntity> userRoles = userdto.getRoles().stream()
+        List<UserRoleEntity> userRoles = userDto.getRoles().stream()
                                                 .map(role -> {
                                                     UserRoleEntity userRole = new UserRoleEntity();
                                                     userRole.setUsername(user.getUsername());
@@ -62,21 +128,28 @@ public class UserServiceImpl implements IUserService {
         user.setRoles(userRoles);
         return userDao.save(user);
     }
+
+    //Metodo para actualizar un usuario por sys_admin
     @Transactional
     @Override
     public UserEntity updateUser(UserUpdateRequestDto userdto) {
 
+        if(!existUsername(userdto.getUsername()) || findById(userdto.getId()) != null){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El usuario no esta registrado" );
+        }
+
         UserEntity user = UserEntity.builder()
+                         .idUser(userdto.getId())
                          .name(userdto.getName())
                          .lastName(userdto.getLastName())
                          .email(userdto.getEmail())
-                         .password(userdto.getPassword())
                          .username(userdto.getUsername())
                          .build();
         return userDao.save(user);
     }
 
 
+    //Metodo para encontrar un usuario por id
     @Transactional
     @Override
     public UserEntity findById(Integer id) {
@@ -84,12 +157,43 @@ public class UserServiceImpl implements IUserService {
         
     }
 
+    //Metodo para eliminar un usuario y da permiso de acuerdo a los roles
     @Transactional
     @Override
-    public void delete(UserEntity user) {
-        userDao.delete(user);
+    public void delete(Integer id) {
+        UserEntity currentUser = findByUsername(getCurrentUsername());
+        UserEntity user = findById(id);
+        if (user==null || currentUser==null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El usuario no esta registrado para ser eliminado" );
+        }
+        String currentUserRole = setDeletePrivileges(currentUser.getRoles());
+        String targetUserRole = setDeletePrivileges(user.getRoles());
+        if(currentUserRole=="ADMIN_SYSTEM"){
+            userDao.delete(user);
+        }
+
+        if(setDeletePrivileges(user.getRoles())=="ADMIN_SYSTEM"){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No se puede eliminar" );
+        }else if(targetUserRole=="ADMIN" && currentUserRole=="OWNER" ){
+            if(user.getStore()==currentUser.getStore()){
+                userDao.delete(user);
+                return;
+            }else{
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No se puede eliminar" );
+            }
+        }else if(targetUserRole=="USER"){
+            if(user.getStore()==currentUser.getStore()){
+                userDao.delete(user);
+                return;
+            }else{
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No se puede eliminar" );
+            }
+        }else{
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No se puede eliminar" );
+        }
     }
 
+    //Metodo para verificar si un usuario existe por su username
     @Transactional
     @Override
     public boolean existUsername(String username){
@@ -100,12 +204,58 @@ public class UserServiceImpl implements IUserService {
         }
     }
 
+    //Metodo para encontrar un usuario por su username
     @Override
     public UserEntity findByUsername(String username) {
          UserEntity user = userDao.findByUsername(username).orElseThrow(null);
         return user;
     }
 
+    //Metodo para obtener la paginacion de usuarios
+    @Override
+    public Page<UserEntity> getUsersByStore( Pageable page) {
+        UserEntity user = findByUsername(getCurrentUsername());
+        return userDao.findAllByStore(user.getStore(), page);
+    }
+
+    @Override
+    public String setPosition(List<UserRoleEntity> roles){
+        for(UserRoleEntity rol: roles){
+            if(rol.getRole() == "OWNER" || rol.getRole() == "ADMIN"){
+                return "ADMIN";
+            }else{
+                break;
+            }
+        }
+        return "Trabajador";
+    }
+
+    //Metodo que devuelve el usuario actual de la peticion
+    public String getCurrentUsername() {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal instanceof UserDetails) {
+            return ((UserDetails) principal).getUsername();
+        } else {
+            return principal.toString();
+        }
+    }
+
+    //Metodo que devuelve el rol dominante 
+    public String setDeletePrivileges(List<UserRoleEntity> roles){
+        List<String> roleNames = new ArrayList<>();
+
+        for(UserRoleEntity rol: roles){
+            roleNames.add(rol.getRole());
+        }
+        if(roleNames.contains("ADMIN_SYSTEM")){
+            return "ADMYN_SYSTEM";
+        }else if(roleNames.contains("OWNER")){
+            return "OWNER";
+        }else if (roleNames.contains("ADMIN")){
+            return "ADMIN";
+        }
+        return "USER";
+    }
 
 
 }
